chore: move nested message_payload module to a dedicated file
